split cmd handlers into files
